Add pluggable backend: support both Sportmonks and API-Football
Split request_handler.py into three files:
- api_football_handler.py  ApiFootballHandler class (API-Football v3)
- sportmonks_handler.py    SportmonksHandler class (Sportmonks V3)
- request_handler.py       factory function — reads backend from config

Switch backends by setting backend = api-football or backend = sportmonks
in the [auth] section of config.ini. New configs default to api-football.
Existing configs without the key are silently upgraded to api-football.
